5-deployment/matchagen/pyproject.toml
[project]
name = "matchagen"
version = "0.1.0"
description = "Matcha recipe generator using distilgpt2"
requires-python = ">=3.11"
dependencies = [
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "torch>=2.0.0",
    "accelerate>=0.26.0",
    "loguru>=0.7.0",
    "tomli>=2.0.0",
    "firecrawl-py>=4.8.0",
    "sentencepiece>=0.1.99",
]

[project.optional-dependencies]
backend = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
